Consolidate external import tests
